Fix old methods eating new methods
At some point, I needed the +1 to capture the }. I can't remember exactly
why, it might have been before I added error checking. Either way it
doesn't seem to be needed now, and it ends up grabbing \r. This makes it
so that if you place your cursor immediately after } and press return,
anything you type will be in the tracking span for the previous method.